data layer POC
