+++ src/index.tsx
@@ -6,20 +6,6 @@ import { ResponsesProvider } from './contexts/responses';
 import { PersonalityProvider } from './contexts/personality';
 import { SessionProvider } from './contexts/session';
 import { ClimateFeedProvider } from './contexts/climateFeed';
-import TagManager from 'react-gtm-module';
-
-const tagManagerArgs = {
-  gtmId: 'GTM-56GRWXW',
-  dataLayer: {
-    userId: '001',
-    userProject: 'Climate Mind App',
-    events: {
-      sendQuestionNumber: 'question_start',
-    },
-  },
-};
-
-TagManager.initialize(tagManagerArgs);
 
 ReactDOM.render(
   <React.StrictMode>
